handle for darwin/mac os

+++ nodejs-installer.sh
@@ -28,39 +28,52 @@ if [[ $VERSION =~ $IS_SPECIFIC ]] ; then
   else
   case $VERSION in
     lts )
-    VERSION='v4.4.7' ;;
+    VERSION='v12.14.1' ;;
     latest )
-    VERSION='v6.4.0' ;;
+    VERSION='v13.7.0' ;;
     *) 
-    VERSION='v4.4.7' ;;
+    VERSION='v12.14.1' ;;
   esac
 fi
-echo "removing old node..."
-sudo apt-get --purge remove node
-sudo apt-get --purge remove nodejs-legacy
-sudo apt-get --purge remove nodejs
-echo "old node remove done!!"
-echo "Changing owner of the folder"
-sudo chown $(whoami):$(whoami) /opt/
-cd /opt
+if [[ "$OSTYPE" == "linux-gnu" ]]; then
+    echo "Removing old node..."
+    sudo apt-get --purge remove node
+    sudo apt-get --purge remove nodejs-legacy
+    sudo apt-get --purge remove nodejs
+    echo "Old node remove done!!"
+fi
+echo "Removing old node gyp and modules"
+rm -rf ~/.node-gyp ~/.npm
+echo "Create Softwares folder"
+mkdir -p ~/Softwares && cd $_
 echo "Downloading node version: $VERSION"
-sleep 3 
-wget "https://nodejs.org/dist/$VERSION/node-$VERSION-linux-x64.tar.gz"
+sleep 2
+if [[ "$OSTYPE" == "darwin"* ]]; then
+    OS_TYPE="darwin"
+else
+    OS_TYPE="linux"
+fi
+DOWNLOAD_FOLDERNAME="node-$VERSION-$OS_TYPE-x64"
+curl -o "$DOWNLOAD_FOLDERNAME.tar.gz" "https://nodejs.org/download/release/$VERSION/node-$VERSION-$OS_TYPE-x64.tar.gz"
 echo "Downloading completed"
-sleep 3
-tar -xvzf node-$VERSION-linux-x64.tar.gz
-rm -rf node
-ln -s node-$VERSION-linux-x64 node
+sleep 2
+tar -xvzf "$DOWNLOAD_FOLDERNAME.tar.gz"
+rm "$DOWNLOAD_FOLDERNAME.tar.gz"
+rm -f node
+ln -s "$DOWNLOAD_FOLDERNAME" node
 echo "Go Back to Home"
 cd ~
-echo "Updating .bashrc file"
-sleep 3
-echo "export NODE=/opt/node" >> .bashrc
-echo 'export PATH=$PATH:$NODE/bin' >> .bashrc
-echo 'export PATH' >> .bashrc
-. .bashrc
-sleep 3
+echo "Updating .bashrc or .bash_profile file"
+sleep 2
+if [[ "$OSTYPE" == "darwin"* ]]; then
+    echo '\nexport NODE=~/Softwares/node\nexport PATH=$PATH:$NODE/bin\nexport PATH' >> .bash_profile
+    source .bash_profile
+else
+    echo '\nexport NODE=~/Softwares/node\nexport PATH=$PATH:$NODE/bin\nexport PATH' >> .bashrc
+    source .bashrc
+fi
+sleep 2
 echo "Installation done!!Please check with"
 echo "node -v"
 echo "Bye! Bye!"
-echo "In case of any error, contact deepak.m.shrma@gmail.com"
+echo "In case of any problem, contact deepak.m.shrma@gmail.com\nGithub:https://github.com/deepakshrma/"
